requirements.txt
accelerate==1.5.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
appdirs==1.4.4
asgiref==3.8.1
attrs==25.1.0
backoff==2.2.1
bcrypt==4.2.1
blinker==1.9.0
boilerpy3==1.0.7
build==1.2.2.post1
cachetools==5.5.1
cattrs==24.1.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
coloredlogs==15.0.1
Deprecated==1.2.18
deprecation==2.1.0
distro==1.9.0
docopt==0.6.2
durationpy==0.9
elastic-transport==8.17.0
elasticsearch==8.11.1
elasticsearch-haystack==2.0.0
Events==0.5
fastapi==0.115.12
filelock==3.17.0
flatbuffers==25.2.10
fsspec==2025.2.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
googleapis-common-protos==1.67.0
grpcio==1.67.1
h11==0.14.0
haystack-ai==2.11.1
haystack-experimental==0.6.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.28.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.5.2
inflect==7.5.0
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kubernetes==32.0.0
lazy-imports==0.3.1
loky==3.5.0
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
more-itertools==10.6.0
mpmath==1.3.0
narwhals==1.26.0
networkx==3.4.2
num2words==0.5.14
numpy==2.2.2
oauthlib==3.2.2
onnxruntime==1.20.1
openai==1.61.1
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-instrumentation==0.51b0
opentelemetry-instrumentation-asgi==0.51b0
opentelemetry-instrumentation-fastapi==0.51b0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
opentelemetry-util-http==0.51b0
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.6
posthog==3.12.1
prompthub-py==4.0.0
protobuf==5.29.3
psutil==7.0.0
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==1.10.21
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
pylance==0.23.0
pypdf==5.3.0
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-env==1.0.0
pytz==2025.1
PyYAML==6.0.2
quantulum3==0.9.2
rank-bm25==0.2.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-cache==0.9.8
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
sentence-transformers==3.4.1
setuptools==75.8.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sseclient-py==1.8.0
starlette==0.45.3
streamlit==1.42.2
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.9.0
tokenizers==0.21.0
toml==0.10.2
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
transformers==4.48.3
typeguard==4.4.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
ujson==5.10.0
url-normalize==1.4.3
urllib3==1.26.20
uvicorn==0.34.0
uvloop==0.21.0
watchdog==6.0.0
watchfiles==1.0.4
websocket-client==1.8.0
websockets==14.2
wrapt==1.17.2
zipp==3.21.0
